DQUERY
Solution for DQUERY problem using BIT tree.. Idea is to first sort the queries in order of their R part. Now we only need to check how many elements in [L,R] are useful for us. On every step we can keep track if the new element was already there or not. If it was there then just update the BIT tree by adding (-1) and thus decreasing the effect.